ui(header) rename to AppHeader
