+++ packages/commercetools/theme/components/AppHeader.vue
